LayoutTests/imported/w3c/web-platform-tests/css/css-animations/historical-expected.txt


FAIL Historical CSS features must be removed: WebKitAnimationEvent assert_equals: expected (undefined) undefined but got (function) function "function WebKitAnimationEvent() {
    [native code]
}"
